refactoring garbage collector
